revert spme pragmas

+++ Doom3/neo/framework/Common.cpp
@@ -102,7 +102,7 @@ bool			com_editorActive;		//  true if an editor has focus
 #ifdef _WIN32
 HWND			com_hwndMsg = NULL;
 bool			com_outputMsg = false;
-unsigned int	com_msgID = -1;
+unsigned int	com_msgID = static_cast<unsigned int>( -1 );
 #endif
 
 #ifdef __DOOM_DLL__
